+++ CHANGELOG.md
@@ -10,6 +10,7 @@ A new section Unreleased is opened then for next dev phase.
 ### Added
 - Add pytest tests. [#23]
 - Add sphinx doc and readthedocs. [#30]
+- Clarifications and make optional input parameter georef. [#71]
 
 ### Changed
 - Limit decimal number on output offsets. [#62]